Unifies map selectors
